requirements.txt
certifi==2019.3.9
chardet==3.0.4
clearbit==0.1.7
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
get==2019.3.22
idna==2.8
pkg-resources==0.0.0
post==2019.3.22
public==2019.3.22
PyJWT==1.7.1
pytz==2018.9
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
urllib3==1.24.1
